services/discovery/requirements.txt
pydantic==2.*
google-cloud-firestore==2.*
fastapi==0.*